Moves HistoryService to separate file
