+++ src/index.ts
@@ -21,22 +21,25 @@ let cell = cells()
 // Handle left and right arrow key presses
 document.removeEventListener('keydown', cell.state.onkeydown)
 document.addEventListener('keydown', function (event) {
-  switch(event.key) {
+  switch (event.key) {
     case 'ArrowLeft':
       cell.update({ historyIndex: cell.getState().historyIndex - 1 })
+      event.preventDefault()
       break
     case 'ArrowRight':
       cell.update({ historyIndex: cell.getState().historyIndex + 1 })
+      event.preventDefault()
       break
     case 'ArrowUp':
       cell.update({ historyIndex: 0 })
+      event.preventDefault()
       break
     case 'ArrowDown':
       cell.update({ historyIndex: cell.getState().history.length })
+      event.preventDefault()
       break
   }
 })
-
 
 // Debug
