+++ src/meiosis.css
@@ -2,7 +2,7 @@
 :root {
     --primary: #7b00ff;
     --secondary: #6c757d;
-    --highlight: rgb(149, 149, 149);
+    --highlight: #adadad;
     --success: #28a745;
     --info: #17a2b8;
     --warning: #ffc107;
